Creation verifying links
