+++ frontend/src/components/organization/EditOrganizationRoot.js
@@ -16,6 +16,7 @@ import UserContext from "../context/UserContext";
 import PageNotFound from "../general/PageNotFound";
 import TranslateTexts from "../general/TranslateTexts";
 import { parseOrganization } from "../../../public/lib/organizationOperations";
+import { verifySocialMediaLinks } from "../../../public/lib/socialMediaOperations";
 
 const useStyles = makeStyles((theme) => ({
   headline: {
@@ -285,50 +286,4 @@ async function getOrganizationByUrlIfExists(organizationUrl, token, locale) {
     if (err.response && err.response.data) console.log("Error: " + err.response.data.detail);
     return null;
   }
-}
-
-function verifySocialMediaLinks(socialOptions, texts) {
-  const err = {};
-
-  // matches http://, https:// , https://www. ,http://www.
-  const regexPrefix = "^(http)(?:s)?(://)(?:www.)?";
-
-  // matches.com/ anything
-  const regexSuffix = ".com/.+$";
-  let regex;
-  let matches = false;
-
-  socialOptions.map((so) => {
-    switch (so.key) {
-      case 0: // twitter
-        regex = new RegExp(regexPrefix + "twitter" + regexSuffix);
-        matches = regex.test(so.social_media_name);
-        err.twitterErr = matches ? null : texts.does_not_comply_twitter;
-        break;
-      case 1: // youtube
-        regex = new RegExp(regexPrefix + "youtube" + regexSuffix);
-        matches = regex.test(so.social_media_name);
-        err.youtubeErr = matches ? null : texts.does_not_comply_youtube;
-        break;
-      case 2: // linkedIn
-        regex = new RegExp(regexPrefix + "linkedin" + regexSuffix);
-        matches = regex.test(so.social_media_name);
-        err.linkedInErr = matches ? null : texts.does_not_comply_linkedin;
-        break;
-      case 3: // instagram
-        regex = new RegExp(regexPrefix + "instagram" + regexSuffix);
-        matches = regex.test(so.social_media_name);
-        err.instagramErr = matches ? null : texts.does_not_comply_instagram;
-        break;
-      case 4: // facebook
-        regex = new RegExp(regexPrefix + "facebook" + regexSuffix);
-        matches = regex.test(so.social_media_name);
-        err.facebookErr = matches ? null : texts.does_not_comply_facebook;
-        break;
-      default:
-        break;
-    }
-  });
-
-  return err;
 }
